Change suite names to avoid use of 'default'.
For reasons that are still unknown to me, when trying to converge the
`default` suite using kitchen-dokken, cinc-solo fails with a 404 trying
to read the `test` environment. This is in spite of the removal of
explicit test arguments and in spite of the fact that suites named
anything other than `default` work fine. It also wasn't a problem for
kitchen-ec2.

This uses a prefix and descriptive name for each test.